moved packages outside of the cmd directory
